fix(refactoring): create GameViewController+GameSceneDelegate
